% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc_path.R
\name{acc_path}
\alias{acc_path}
\title{Generate a shortest path raster from a specific graph node or set of graph coordinates}
\usage{
acc_path(graph, costsurf, snode = NULL, scoord = NULL)
}
\arguments{
\item{graph}{igraph object}

\item{costsurf}{RasterLayer cost surface}

\item{snode}{numeric starting node number}

\item{scoord}{matrix 2 column starting coordinate}
}
\description{
The shortest.paths function from the igraph package is applied to an igraph object and the results are coerced to a raster object.
}

