% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdist_acc.R
\name{gdist_acc}
\alias{gdist_acc}
\title{Generate an accumulated cost surface using a regular landscape graph and the gdistance transition class}
\usage{
gdist_acc(costsurf, scoord = NULL, snode = NULL, geocorrect = TRUE)
}
\arguments{
\item{costsurf}{Raster}

\item{scoord}{matrix 2 column}

\item{snode}{numeric starting node number}

\item{geocorrect}{logical use gdistance::geoCorrection() on transition matrix?}
}
\description{
Generate an accumulated cost surface using a regular landscape graph and the gdistance transition class
}
\details{
need to add ability to call gdistance::geoCorrection
}
\examples{
dm <- as.matrix(read.delim(system.file("extdata/etherington20120code/cost-surface20x20.txt",
package = "irlgraph"), skip = 6, na.strings = "-9999", header = FALSE, sep = " "))

costsurf <- raster::raster(nrows=dim(dm)[1],ncols=dim(dm)[2],
resolution=1,xmn=0, xmx = dim(dm)[1], ymn = 0, ymx = dim(dm)[2])
#neccessary to set resolution

costsurf[] <- dm
scoord <- matrix(c(10.5, 10.5, 15.5, 15.5), ncol = 2, byrow = TRUE)
result <-  gdist_acc(costsurf, scoord)

sp::plot(result[[1]]$result)
points(scoord[1,1], scoord[1,2])

sp::plot(result[[2]]$result)
points(scoord[2,1], scoord[2,2])
}

