% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{Generate a shortest path raster from a specific graph node}
\usage{
get_path(egraph2, snode, csurf)
}
\description{
Generate a shortest path raster from a specific graph node
}

