\name{ilgraph-package}
\alias{ilgraph-package}
\alias{ilgraph}
\docType{package}
\title{
\packageTitle{ilgraph}
}
\description{
\packageDescription{ilgraph}
}

\author{
\packageAuthor{ilgraph}

Maintainer: \packageMaintainer{ilgraph}
}
\references{
Etherington, T. R. (2012). Least-cost modelling on irregular landscape graphs. Landscape ecology, 27(7), 957-968.
}

\keyword{ package }