% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Impute NA cells according to nearest neighbor}
\usage{
impute_na(csurf, cells, nullcells, result, cellcoords)
}
\arguments{
\item{csurf}{}

\item{cells}{}

\item{nullcells}{}

\item{result}{}

\item{cellcoords}{}
}
\description{
Impute NA cells according to nearest neighbor
}

