% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irgraph.R
\name{irgraph}
\alias{irgraph}
\title{Generate an irregular landscape graph}
\usage{
irgraph(dm)
}
\arguments{
\item{dm}{matrix cost raster}
}
\description{
Generate an irregular landscape graph
}
\examples{
dm<-matrix(c( 1,  1,  1,  1,  1,
 NA, NA, NA, NA, 1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1
 ), ncol = 5, byrow = TRUE)

irgraph(dm)
}

