% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irl_acc.R
\name{irl_acc}
\alias{irl_acc}
\title{Generate an accumulated cost surface from an irregular landscape graph}
\usage{
irl_acc(dm, poicoords = NA, cutoff = 0, grainprop = 0.25, costsurf,
  scoord, snode = NULL, irregular = TRUE, warn = TRUE)
}
\arguments{
\item{dm}{matrix cost raster}

\item{poicoords}{matrix}

\item{cutoff}{numeric between 0 and 1. Determines the threshold at which a cell is determined "very important".}

\item{grainprop}{numeric proportion of cells to be included as "grain cells".}

\item{costsurf}{Raster cost surface}

\item{scoord}{matrix specifying starting coordinates. Must be 2 columns.}

\item{snode}{numeric starting node}

\item{irregular}{logical}

\item{warn}{logical}
}
\description{
This high level function allows for generation of an accumulated cost surface from a single function call.
}
\examples{
\dontrun{
set.seed(123) #make reproducible

dm <- as.matrix(read.delim(system.file(
"extdata/etherington20120code/cost-surface20x20.txt",
package = "irlgraph"), skip = 6, na.strings = "-9999", header = FALSE, sep = " "))

costsurf <- raster::raster(nrows=dim(dm)[1],ncols=dim(dm)[2],resolution=1,xmn=0,
xmx = dim(dm)[1], ymn = 0, ymx = dim(dm)[2]) #neccessary to set resolution
costsurf[] <- dm

poicoords <- matrix(c(10.5, 10.5), ncol = 2)

result <- irl_acc(dm, poicoords = poicoords, grainprop = 0.25,
costsurf = costsurf, scoord = matrix(c(19.5, 0.5, 11.5, 11.5),
ncol = 2, byrow = TRUE))

result <- irl_acc(dm, poicoords = poicoords, grainprop = 0.25,
costsurf = costsurf, scoord = matrix(c(11.5, 11.5),
ncol = 2, byrow = TRUE))
}
}

