% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irgraph.R
\name{irl_graph}
\alias{irl_graph}
\title{Generate an irregular landscape graph}
\usage{
irl_graph(dm, poicoords = NA, grainprop = 0.25)
}
\arguments{
\item{dm}{matrix cost raster}

\item{poicoords}{matrix}

\item{grainprop}{numeric}
}
\description{
Generate an irregular landscape graph
}
\examples{
dm<-matrix(c( 1,  1,  1,  1,  1,
 NA, NA, NA, NA, 1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1
 ), ncol = 5, byrow = TRUE)

ir_graph(dm)
}

