\name{irlgraph-package}
\alias{irlgraph-package}
\alias{irlgraph}
\docType{package}
\title{
\packageTitle{irlgraph}
}
\description{
\packageDescription{irlgraph}
}

\author{
\packageAuthor{irlgraph}

Maintainer: \packageMaintainer{irlgraph}
}
\references{
Etherington, T. R. (2012). Least-cost modelling on irregular landscape graphs. Landscape ecology, 27(7), 957-968.
}

\keyword{ package }