% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_path.R
\name{acc_path}
\alias{acc_path}
\title{Generate a shortest path raster from a specific graph node}
\usage{
acc_path(graph, snode = NULL, scoord = NULL, costsurf)
}
\arguments{
\item{snode}{}

\item{scoord}{}

\item{egraph2}{}

\item{csurf}{}
}
\description{
Generate a shortest path raster from a specific graph node
}

