% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdist_acc.R
\name{gdist_acc}
\alias{gdist_acc}
\title{Generate an accumulated cost surface using a regular landscape graph and the gdistance transition class}
\usage{
gdist_acc(costsurf, scoord = NULL, snode = NULL)
}
\arguments{
\item{costsurf}{}

\item{scoord}{}

\item{snode}{}
}
\description{
Generate an accumulated cost surface using a regular landscape graph and the gdistance transition class
}
\examples{
dm <- as.matrix(read.delim(system.file("extdata/etherington20120code/cost-surface20x20.txt", package = "irlgraph"),
skip = 6, na.strings = "-9999", header = FALSE, sep = " "))

costsurf <- raster::raster(nrows=dim(dm)[1],ncols=dim(dm)[2],resolution=1,xmn=0, xmx = dim(dm)[1], ymn = 0, ymx = dim(dm)[2]) #neccessary to set resolution

costsurf[] <- dm
scoord <- c(10.5, 10.5)
plot(gdist_acc(costsurf, scoord))
}

