% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Impute NA cells according to nearest neighbor}
\usage{
impute_na(result, csurf, graph, warn = TRUE)
}
\arguments{
\item{result}{Raster}

\item{csurf}{Raster cost surface}

\item{graph}{igraph object}

\item{warn}{logical print warnings?}
}

