% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irl_graph.R
\name{irl_graph}
\alias{irl_graph}
\title{Generate an irregular landscape graph}
\usage{
irl_graph(dm, poicoords = NA, grainprop = 0.25, cutoff = 0,
  irregular = TRUE)
}
\arguments{
\item{dm}{matrix of 2d cost values}

\item{poicoords}{matrix 2 column point-of-interest coordinates}

\item{grainprop}{numeric proportion of cells selected as grain cells}

\item{cutoff}{numeric user-defined importance cutoff}

\item{irregular}{logical skip irregular node subsetting routine?}
}
\description{
Subset the nodes in a matrix of cost values based on user defined point-of-interest coordinates, a user-defined importance cutoff, and a randomly selected proportion of non-null cells. Generate a graph object from this subset using delaunay triangulation.
}
\examples{
\dontrun{

dm <- matrix(
c(1,  1,  1,  1,  1,
 NA, NA, NA, NA, 1, 
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1
 ), ncol = 5, byrow = TRUE)
 
sp::plot(raster::raster(dm))
raster::text(raster::raster(dm), which(raster::raster(dm)[] == 1))

graph <- irl_graph(dm)
plot(graph$graph)
geometry::trimesh(graph$tri$tri, graph$allcoords)
}
}

