% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irl_graph.R
\name{irl_graph}
\alias{irl_graph}
\title{Generate an irregular landscape graph}
\usage{
irl_graph(dm, poicoords = NA, grainprop = 0.25, cutoff = 0,
  irregular = TRUE)
}
\arguments{
\item{dm}{matrix cost raster}

\item{poicoords}{matrix}

\item{grainprop}{numeric}

\item{cutoff}{numeric}

\item{irregular}{logical}
}
\description{
Generate an irregular landscape graph
}
\examples{
dm<-matrix(c( 1,  1,  1,  1,  1,
 NA, NA, NA, NA, 1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1,
 1,  1,  1,  1,  1
 ), ncol = 5, byrow = TRUE)

sp::plot(raster::raster(dm))
raster::text(raster::raster(dm), which(raster::raster(dm)[] == 1))

graph <- irl_graph(dm)
plot(graph$graph)
geometry::trimesh(graph$tri$tri, graph$allcoords)
}

